+++ common/src/main/java/org/opensearch/ml/common/connector/HttpConnector.java
@@ -32,7 +32,7 @@ import java.util.regex.Pattern;
 import static org.opensearch.core.xcontent.XContentParserUtils.ensureExpectedToken;
 import static org.opensearch.ml.common.connector.ConnectorProtocols.HTTP;
 import static org.opensearch.ml.common.connector.ConnectorProtocols.validateProtocol;
-import static org.opensearch.ml.common.input.Constants.TENANT_ID;
+import static org.opensearch.ml.common.CommonValue.TENANT_ID;
 import static org.opensearch.ml.common.utils.StringUtils.getParameterMap;
 import static org.opensearch.ml.common.utils.StringUtils.isJson;
 import org.opensearch.ml.common.transport.connector.MLCreateConnectorInput;
